requirements.txt
# Core Framework
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0

# Security
Werkzeug==3.0.1
python-dotenv==1.0.0

# External API calls
requests==2.31.0

# CORS support (useful for testing with web interfaces)
Flask-CORS==4.0.0

# Swagger/OpenAPI documentation
flask-swagger-ui==4.11.1